requirements.txt
amqp==2.2.2
billiard==3.5.0.3
bleach==1.5.0
certifi==2017.7.27.1
chardet==3.0.4
coreapi==2.3.1
coreschema==0.0.4
Django==1.11.5
django-cors-headers==2.1.0
djangorestframework==3.6.4
djangorestframework-csv==2.0.0
djangorestframework-jwt==1.11.0
enum34==1.1.6
html5lib==0.9999999
idna==2.6
itypes==1.1.0
Jinja2==2.9.6
jsonfield==2.0.2
kombu==4.1.0
Markdown==2.6.10
MarkupSafe==1.0
numpy==1.13.3
protobuf==3.5.1
psutil==5.4.2
PyJWT==1.5.3
pytz==2017.2
requests==2.18.4
scikit-learn==0.19.1
scipy==1.0.0
six==1.11.0
sklearn==0.0
tensorflow==1.4.1
tensorflow-tensorboard==0.4.0rc3
unicodecsv==0.14.1
uritemplate==3.0.0
urllib3==1.22
views==0.3
vine==1.1.4
Werkzeug==0.13
